stop autothemeprovider from leaking dependencies
