+++ packages/themes/package.json
@@ -14,6 +14,7 @@
     "registry": "https://registry.npmjs.org/",
     "access": "public"
   },
+  "sideEffects": false,
   "scripts": {
     "clean": "ds clean",
     "build": "ds build",
