+++ packages/themes/src/index.tsx
@@ -1,2 +1,3 @@
 export * from "./themes";
 export * from "./utils";
+export * from "./AutoThemeProvider";
