+++ packages/themes/src/utils.tsx
@@ -1,8 +1,6 @@
 import React from "react";
 import makeClass from "clsx";
-import { isFirefox } from "react-device-detect";
 import { Element } from "@design-systems/utils";
-import { all } from "./themes";
 
 export const colorSchemes = ["light", "dark"] as const;
 export type ColorScheme = typeof colorSchemes[number];
@@ -75,44 +73,6 @@ export const ThemeProvider = ({
   return (
     <ThemeContext.Provider value={{ ...wrappedTheme, ...value }}>
       {children}
-    </ThemeContext.Provider>
-  );
-};
-
-export interface AutoThemeProviderProps extends ThemeableElement<"div"> {
-  /** Whether to automatically change the font and background color */
-  autoStyle?: boolean;
-  /** Any React node children */
-  children: React.ReactNode;
-}
-
-/**
- * A theme provider that automatically detects a users browser and colorScheme.
- * Themes are set for each component using React Context.
- * It also sets the background color and text color to the correct color.
- */
-export const AutoThemeProvider = ({
-  theme: propsTheme,
-  colorScheme: propsColorScheme,
-  autoStyle,
-  children,
-  ...html
-}: AutoThemeProviderProps) => {
-  const isDark = useDarkMode();
-  const colorScheme = propsColorScheme || (isDark ? "dark" : "light");
-  const theme = propsTheme || (isFirefox ? "firefox" : "chrome");
-  const style = {
-    backgroundColor: all[theme][colorScheme].backgroundColor,
-    color: all[theme][colorScheme].textColor,
-    minHeight: "100%",
-    width: "100%",
-  };
-
-  return (
-    <ThemeContext.Provider value={{ theme, colorScheme }}>
-      <div style={autoStyle ? style : undefined} {...html}>
-        {children}
-      </div>
     </ThemeContext.Provider>
   );
 };
